add leaderboard

+++ src/cpp/Commands/Double.cpp
@@ -16,7 +16,7 @@ Double::Double() {
 // Execute method
 void Double::executeCommand(CandyGameState& gameState) {
     unsigned long long oldPoints = gameState.getPointPool();
-    unsigned long long newPoints = oldPoints * 2;
+    unsigned long long newPoints = oldPoints * (unsigned long long)2;
     gameState.setPointPool(newPoints);
 
     cout << gameState.getCurrentTurnPlayer().getName() << " melakukan DOUBLE! Poin hadiah naik dari " << oldPoints << " menjadi " << newPoints << "!" << endl;
